requirements.txt
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
requests>=2.31
beautifulsoup4>=4.12
scrapy>=2.11
pandas>=2.2
python-dotenv>=1.0
rapidfuzz>=3.6
fastapi>=0.110
uvicorn>=0.29
httpx>=0.27
soccerdata>=1.7
pytest>=8.1
